Game logic is closer to being done
Game logic now uses the GameState object.  Most of the individual
functions in the game logic are there now, but the major ones left to
add are scoring and combo/chain detecting.  After that, the game loop
has to be tested to make sure everything's working in the right order.